feat(hunger): add HungerComponent and wire starvation damage
- HungerComponent: current/maximum/percentage/is_starving, update/consume/reset
- PlayerData: max_hunger, hunger_decay_rate, hunger_damage_rate (all from JSON)
- Player.update(): orchestrates hunger tick → starvation → take_damage
- HealthComponent/Entity: take_damage/heal accept float (fractional per-frame damage)
- 12 new tests (10 unit + 2 integration); total 58/58 green, 86% coverage

+++ core/entity.py
@@ -15,7 +15,7 @@ class Entity(GameObject):
         """Жива ли сущность."""
         return self.health.is_alive
 
-    def take_damage(self, amount: int) -> None:
+    def take_damage(self, amount: float) -> None:
         """Нанести урон. Делегирует компоненту, затем оркестрирует события."""
         self.health.take_damage(amount)
         if not self.is_alive:
@@ -24,7 +24,7 @@ class Entity(GameObject):
         else:
             EventBus.emit("entity_damaged", {"entity": self, "amount": amount})
 
-    def heal(self, amount: int) -> None:
+    def heal(self, amount: float) -> None:
         """Восстановить HP, не превышая максимум."""
         self.health.heal(amount)
         EventBus.emit("entity_healed", {"entity": self, "amount": amount})
