+++ systems/health.py
@@ -1,17 +1,17 @@
 class HealthComponent:
     """Компонент здоровья. Хранит HP и предоставляет логику изменений без побочных эффектов."""
 
-    def __init__(self, maximum: int) -> None:
-        self._maximum: int = maximum
-        self._current: int = maximum
+    def __init__(self, maximum: float) -> None:
+        self._maximum: float = maximum
+        self._current: float = maximum
 
     @property
-    def current(self) -> int:
+    def current(self) -> float:
         """Текущее здоровье."""
         return self._current
 
     @property
-    def maximum(self) -> int:
+    def maximum(self) -> float:
         """Максимальное здоровье."""
         return self._maximum
 
@@ -25,11 +25,11 @@ class HealthComponent:
         """Доля HP от максимума в диапазоне 0.0–1.0. Используется для HP-бара."""
         return self._current / self._maximum if self._maximum > 0 else 0.0
 
-    def take_damage(self, amount: int) -> None:
+    def take_damage(self, amount: float) -> None:
         """Уменьшить HP на amount, не опускаясь ниже нуля."""
-        self._current = max(0, self._current - amount)
+        self._current = max(0.0, self._current - amount)
 
-    def heal(self, amount: int) -> None:
+    def heal(self, amount: float) -> None:
         """Восстановить HP на amount, не превышая максимум."""
         self._current = min(self._maximum, self._current + amount)
